Fix error when data was being loaded while logout was in progress
The Routes component skips calling the possible loadData of a page if
the logout is in progress. However, the logoutSuccess and the
clearCurrentUser actions were dispatched in an order that cleared the
user from the store before the logout was marked done.

The fix was to order the actions so that the logout is marked done
first and then all the cleanup is done.

A better solution in the future would be to batch actions so that
extra render loops are avoided when dispatching multiple actions at
once.